requirements.txt
Django==4.0.6
django-model-utils==4.2.0
python-decouple==3.6
